Implement simple KeyValues binary serialization. Includes unit test.
Since KeyValues only currently works with string values, the serialization will write either:
1. An object (.None, name, children, .End)
2. A string (.String, name, value)

Due to the way the deserializer works, serialization and deserialization is not symmetric. Serializing a KV object to binary and then
deserializing it will yield a new KV object, with the original KV object as the sole child of the new one.