Extend default EClientPersonaStateFlag info
Default value that steam client uses right now is equal to ```3154 = PlayerName, Presence, LastSeen, ClanTag, Facebook```. While I'm sure we don't really need ```ClanTag``` or ```Facebook``` info, ```LastSeen``` can be actually pretty useful, for example in my case when I'd like to remove friends that didn't log in since X time. This would allow me as well as every interested person to skip asking for that extra info in another request.

Since this request is default in steam client, and it results in generation of barely 2 extra fields (```LastLogOn``` and ```LastLogOff```), I think this should be added to default info.