Fix machine_id on windows to be like the steam client
First, the first available physical network interface is found. Using registry keys to differentiate the physical ones.

Then, instead of hashing only the MAC address, we hash 16 bytes where the first 8 are the MAC address.

That is because the steam client on windows hashes the first 16 bytes of a struct containing the mac address at the start and only zeroes afterwards.