Handle CMClient.OnClientConnected() throwing (probably from a derived class)
When playing around with NativeAOT, this can easily throw an exception as this
is the first point at which we touch protobuf serialization, particularly when
working with websockets.

At the moment this exception gets captured in the WebSocketContext task and does
not get printed to DebugLog, so the program has the appearance of hanging.

This prints the exception to DebugLog and then triggers a non-user-initiated
Disconnect call to close the connection, rather than just hanging forever on an
open connection that we cannot do anything with.